Some PPU fixed

+++ src/ppu.h
@@ -150,7 +150,7 @@ class PPU {
     // $2006 PPUADDR
     // $2007 PPUDATA
 
-    uint16_t oamaddr;     ///< 8-bit (?) OAM address register
+    uint8_t oamaddr;      ///< 8-bit (?) OAM address register
     uint8_t oamdata;      ///< 8-bit OAM data buffer
     uint8_t ppudata_buf;  ///< 8-bit PPUADDR read buffer
